api/requirements.txt
Django
psycopg2-binary
gunicorn
whitenoise
djangorestframework
django-cors-headers